Move AdsSetupCTAWidget.
